+++ changelog.md
@@ -2,10 +2,10 @@
 - double check :when updating invoice details, the corresponging filtered pages should also be updated. 
 - cold loading client page, e.g. http://localhost:3000/client-dashboard?id=18e50f68-0b4c-4c2a-9bdf-5fe9dcedf3ae
 - clients should be also sortable and paginable
-- add context menu to invoice table
 - all clients should be loaded, otherwise client selection list all options.
 - Refactor/organize based on [slide for Week 5 Office Hour 1](https://docs.google.com/presentation/d/1pLZ8Zvpyr_myotpvyVYzY5awpUFUO78DE5CnBnVNm7c/edit#slide=id.g117828bad90_0_9)
 - throwing exception here: invoiceslice: thunkAPI.dispatch(beforeUpdate(prevInvoice)); (missing client)
+- latest client invoices in profile (link menu item from client table)
 - if you are in page 10, and filter by client, filters are applied but page is still 10 (unexistant)
 
 # extra:
@@ -18,6 +18,7 @@
 # 2022 03 20
 - cancel buttons should go back to previous page
 - Reeplaced hamburger custom menu with MUI PopupState Menu https://mui.com/components/menus/#popupstate-helper
+- add context menu to invoice table
 # 2022 03 19
 - Add support in invoice store slice to filter and sort invoices.
 - Show more columns in the invoices pages